requirements.txt
instructor==1.3.4
openai==1.35.10
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1